# WearBloomContract/Sources/WearBloomContract/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WearBloom API",
    "version": "1.0.0",
    "description": "Private, authenticated API for the WearBloom iOS app."
  },
  "servers": [
    {
      "url": "https://api.wearbloom.app"
    }
  ],
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string",
                "format": "uuid"
              }
            },
            "required": [
              "code",
              "message",
              "requestId"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "RenderVariant": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "lookId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "processing",
              "succeeded",
              "failed",
              "cancelled"
            ]
          },
          "resultURL": {
            "type": "string",
            "nullable": true
          },
          "errorCode": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "lookId",
          "status",
          "resultURL",
          "errorCode",
          "createdAt",
          "completedAt"
        ]
      }
    },
...
  },
  "paths": {
    "/v1/attest/challenge": {
      "get": {
        "operationId": "getAttestChallenge",
        "tags": [
          "Integrity"
        ],
        "responses": {
          "200": {
            "description": "One-time App Attest challenge",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "challenge": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "required": [
                    "challenge"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Request integrity rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid business input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/attest/verify": {
      "post": {
        "operationId": "verifyAppAttest",
        "tags": [
          "Integrity"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "challenge": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "keyId": {
                    "type": "string",
                    "minLength": 10
                  },
                  "attestation": {
                    "type": "string",
                    "minLength": 20
                  }
                },
                "required": [
                  "challenge",
                  "keyId",
                  "attestation"
                ]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "App Attest key enrolled"
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Request integrity rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid business input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/uploads": {
      "post": {
        "operationId": "uploadImage",
        "tags": [
          "Images"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-app-attest-challenge",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-app-attest-key-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-app-attest-assertion",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-app-attest-unsupported",
            "in": "header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": {
                    "nullable": true
                  },
                  "purpose": {
                    "type": "string",
                    "enum": [
                      "garment",
                      "reference"
                    ]
                  }
                },
                "required": [
                  "purpose"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Private image stored",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "contentType": {
                      "type": "string"
                    },
                    "width": {
                      "type": "number",
                      "nullable": true
                    },
                    "height": {
                      "type": "number",
                      "nullable": true
                    }
                  },
                  "required": [
                    "id",
                    "contentType",
                    "width",
                    "height"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Request integrity rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid upload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/renders": {
      "post": {
        "operationId": "createRender",
        "tags": [
          "Renders"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-app-attest-challenge",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-app-attest-key-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-app-attest-assertion",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-app-attest-unsupported",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 200
            },
            "required": true,
            "name": "idempotency-key",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "renderId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "lookId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "referencePhotoId": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                "required": [
                  "renderId",
                  "lookId",
                  "referencePhotoId"
                ]
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Render queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenderVariant"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Request integrity rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid business input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Allowance exhausted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}